ExperimentoSeguridad/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
sqlalchemy==1.4.23
# Mantienes python-jose si quieres, pero evita usarlo para encode/decode
python-jose[cryptography]==3.3.0
# Subo PyJWT y cryptography a versiones estables y con wheel precompilado en Windows
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography>=42.0.0
requests==2.32.3

# Dependencias para el experimento
redis==5.0.1
prometheus-client==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
psycopg2-binary==2.9.9

